fix-tvm-integration
