+++ tools/docker/gluon_nlp_job.sh
@@ -23,7 +23,7 @@ git fetch origin $SOURCE_REF:working
 git checkout working
 
 if [ $DEVICE == "cpu" ]; then
-  python3 -m pip install -U --quiet --pre "mxnet>=2.0.0b20201206" -f https://dist.mxnet.io/python --user
+  python3 -m pip install -U --quiet --pre "mxnet>=2.0.0b20210122" -f https://dist.mxnet.io/python --user
 else
   # Due to the issue in https://forums.aws.amazon.com/thread.jspa?messageID=953912
   # We need to manually configure the shm to ensure that Horovod is runnable.
@@ -31,7 +31,7 @@ else
   umount shm
   mount -t tmpfs -o rw,nosuid,nodev,noexec,relatime,size=2G shm /dev/shm
   sudo python3 -m pip uninstall --quiet mxnet-cu102
-  python3 -m pip install -U --quiet --pre "mxnet-cu102>=2.0.0b20201206" -f https://dist.mxnet.io/python/cu102 --user
+  python3 -m pip install -U --quiet --pre "mxnet-cu102>=2.0.0b20210122" -f https://dist.mxnet.io/python/cu102 --user
 fi
 
 python3 -m pip install --quiet -e .[extras]
