+++ tools/docker/ubuntu18.04-gpu.Dockerfile
@@ -32,7 +32,7 @@ RUN bash /install/install_python_packages.sh
 RUN bash /install/install_tvm_gpu.sh
 
 # Install MXNet
-RUN python3 -m pip install -U --pre "mxnet-cu102>=2.0.0b20201206" -f https://dist.mxnet.io/python --user
+RUN python3 -m pip install -U --pre "mxnet-cu102>=2.0.0b20210122" -f https://dist.mxnet.io/python --user
 
 # Install PyTorch
 RUN python3 -m pip install -U torch torchvision --user
